+++ src/ct_datatypes.jl
@@ -61,6 +61,8 @@ abstract type MixedOhmicSchottkyContact end
 
 abstract type SchottkyBarrierLowering end
 
+abstract type GateContact end
+
 ############    inner boundary conditions     ############
 """
 $(TYPEDEF)
@@ -79,7 +81,7 @@ abstract type InterfaceRecombination end
 """
 Possible types of outer boundary model.
 """
-const OuterBoundaryModelType = Union{Type{OhmicContact}, Type{SchottkyContact}, Type{SchottkyBarrierLowering}, Type{MixedOhmicSchottkyContact}}
+const OuterBoundaryModelType = Union{Type{OhmicContact}, Type{SchottkyContact}, Type{SchottkyBarrierLowering}, Type{MixedOhmicSchottkyContact}, Type{GateContact}}
 
 
 """