+++ src/ct_physics.jl
@@ -607,6 +607,43 @@ function breaction!(f, u, bnode, data, ::Type{InterfaceRecombination})
     return
 end
 
+###########################################################################
+###########################################################################
+
+
+"""
+$(TYPEDSIGNATURES)
+Creates boundary conditions for gate contacts. A Robin boundary condition is applied to the electrostatic potential
+    
+``\\varepsilon_\\mathrm{s} \\nabla \\psi \\cdot \\nu + \\frac{\\varepsilon_\\mathrm{ox}}{d_\\mathrm{ox}} (\\psi - U_G) = Q_{ss}``,
+
+where ``\\varepsilon_\\mathrm{ox}`` denotes the absolute dielectric permittivity of the oxide and ``d_\\mathrm{ox}`` the thickness of the oxide.
+The term ``Q_{ss}`` corresponds to the surface charge density at the gate contact.
+
+For the quasi Fermi potentials, homogeneous Neumann boundary conditions are implemented.
+
+Note that an additional reference voltage ``U_{ref}`` can be absorbed into the surface charge term.
+This leads to an effective surface charge density
+
+``Q_{ss}^{'} = Q_{ss} + \\frac{\\varepsilon_\\mathrm{ox}}{d_\\mathrm{ox}} U_{ref}``.
+
+Equivalently, in terms of surface state density,
+
+``Q_{ss}^{'} = q N_{ss}^{'} ``.
+"""
+
+function breaction!(f, u, bnode, data, ::Type{GateContact})
+
+    params = data.params
+    ipsi = data.index_psi
+
+    # Homogeneous Neumann boundary conditions for electrons and holes by default
+    # Robin boundary condition for the electrostatic potential
+    f[ipsi] = (params.dielectricConstantOxideGate[bnode.region] / params.thicknessOxideGate[bnode.region]) * (u[ipsi] - params.contactVoltage[bnode.region]) - params.surfaceChargeDensityGate[bnode.region]
+
+    return
+end
+
 ##########################################################
 ##########################################################